% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchCV.R
\name{benchCV}
\alias{benchCV}
\title{Benchmarking methods with cross-validation.}
\usage{
benchCV(
  rep,
  k = 5,
  y,
  x,
  z,
  y_order = NULL,
  std = TRUE,
  eval_type,
  niter = 100,
  spec = 0.1,
  fdr = 0.1,
  f_size = 100,
  method = 4,
  delta,
  lambda = 0.1,
  thresh_fdr = 0.2,
  rep_cv = 50,
  alpha_level = 0.05,
  parallel = T,
  ncore = 10,
  out_path
)
}
\arguments{
\item{rep}{an integer; the replicate number}

\item{k}{an integer; the number of folds for cross-validation}

\item{y}{a matrix or data frame of responses of dimensions \eqn{n} by 1}

\item{x}{a matrix or data frame of measured data of dimensions \eqn{n} by \eqn{p}}

\item{z}{a matrix or data frame; latent factors identified using \code{x}}

\item{y_order}{a vector; the levels of \code{y} if it is ordinal}

\item{std}{a boolean flag; whether standardization is performed outside (FALSE) or within cross-validation splitting (TRUE)}

\item{eval_type}{a string either "corr" or "auc"; the method by which to evaluate performance}

\item{niter}{an integer; the number of times to run knockoffs}

\item{spec}{a numeric constant; the proportion of iterations that a variable must be selected in to be considered significant}

\item{fdr}{a numeric constant between 0 and 1; the target false discovery rate for knockoffs}

\item{f_size}{an integer; the target size for the subset (number of columns in each subset)}

\item{method}{an integer (1, 2, 3, 4); the selection method for determining marginal/interaction significance}

\item{delta}{\eqn{\delta}, a numerical value used for thresholding}

\item{lambda}{\eqn{\lambda}, a numerical constant used in thresholding}

\item{thresh_fdr}{a numerical constant used for thresholding the correlation matrix to
control the false discovery rate, default is 0.2}

\item{rep_cv}{number of replicates for inner cross-validation of \code{delta}}

\item{alpha_level}{\eqn{\alpha}, a numerical constant used in confidence interval calculation}

\item{parallel}{a boolean flag; run iterations in parallel or sequentially}

\item{ncore}{an integer; the number of cores to use for parallel computing}

\item{out_path}{a string path; where to save output}
}
\value{
a matrix of model performances
}
\description{
Runs one replicate of a cross-validation regime for method performance comparison. Currently supports
plainER, SLIDE, LASSO regression, PHATE regression. Additional methods compared include principal components
regression and partial least squares regression for continuous response and principal components logistic regression
and partial least squares discriminant analysis for categorical response.
}
