% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcDefaultFsize.R
\name{calcDefaultFsize}
\alias{calcDefaultFsize}
\title{calculate the default f_size}
\usage{
calcDefaultFsize(y, all_latent_factors)
}
\arguments{
\item{y}{a matrix of the response vector}

\item{all_latent_factors}{a object contain all latent factors from getLatentFactors.}
}
\value{
a integer representing the default f_size.
}
\description{
calculate the default f_size
}
