% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcZMatrix.R
\name{calcZMatrix}
\alias{calcZMatrix}
\title{calculate Z matrix}
\usage{
calcZMatrix(x_std, all_latent_factors, x_path = NULL, lf_path = NULL, out_path)
}
\arguments{
\item{x_std}{a matrix of input data.}

\item{all_latent_factors}{a object from getLatentFactors.}

\item{x_path}{a string that points to the x matrix.}

\item{lf_path}{a string that points to the getLatentFactors result as an RDS object.}

\item{out_path}{a string that points to the folder where the z matrix will be outputted.}
}
\value{
The z matrix calculated by the PredZ function
}
\description{
Calculate Z matrix by importing the PredZ function from EssReg package
}
