% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkDataParams.R
\name{checkDataParams}
\alias{checkDataParams}
\title{Check if the input and the yaml files are set correctly.}
\usage{
checkDataParams(yaml_path, pipeline = NULL)
}
\arguments{
\item{yaml_path}{a string that points to the yaml file.}

\item{pipeline}{a integer indicates which step of the pipeline is the yaml file for.}
}
\description{
Check if the input and the yaml files are set correctly.
}
