% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkElement.R
\name{checkElement}
\alias{checkElement}
\title{Check Element In List.}
\usage{
checkElement(element, groupList)
}
\arguments{
\item{element}{an element}

\item{groupList}{a list}
}
\value{
group and sublist index if found, \code{(0, 0)} otherwise
}
\description{
Check if an element is in a list. If it does exist in some group, return the group
index in that list and its sublist index. Otherwise, return \code{(0,0)}.
}
