% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compASV.R
\name{compASV}
\alias{compASV}
\title{compASV.}
\usage{
compASV(
  sigma_alt,
  BI,
  Theta_hat,
  Gamma_hat,
  beta_hat,
  Omega_hat,
  I_hat,
  I_hat_list
)
}
\arguments{
\item{sigma_alt}{a matrix computed by \code{\link{estSigmaAlt}}}

\item{BI}{estimated matrix:
\eqn{\hat{A}_{\hat{I}\cdot}(\hat{A}_{\hat{I}\cdot}^\top \hat{A}_{\hat{I}\cdot}^\top)^{-1}}}

\item{Theta_hat}{estimated matrix:
\eqn{(\hat{\Sigma}_{\cdot \hat{I}}-\hat{\Gamma}_{\cdot \hat{I}})BI}}

\item{Gamma_hat}{estimated matrix}

\item{beta_hat}{matrix of the estimated values of \eqn{\beta}}

\item{Omega_hat}{estimated matrix}

\item{I_hat}{vector of pure node indices}

\item{I_hat_list}{list of pure node indices by cluster}
}
\value{
???
}
\description{
I am not sure what this does.
}
