% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correctIt.R
\name{correctIt}
\alias{correctIt}
\title{Correct the response for the marginal when finding interactions.}
\usage{
correctIt(z, y)
}
\arguments{
\item{z}{a vector or data frame with one column of numeric values; the values of a marginal variable}

\item{y}{a vector or data frame with one column of numeric values; the true response}
}
\value{
a vector of numeric values
}
\description{
The new response vector is the residuals after regressing the true response values on the values of
one marginal \eqn{Z}. This is used for determining which interaction terms that include this marginal
variable are truly significant.
}
