% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corrToP.R
\name{corrToP}
\alias{corrToP}
\title{Correlation \eqn{p}-values.}
\usage{
corrToP(r, n)
}
\arguments{
\item{r}{numeric, correlation}

\item{n}{numeric, sample size}
}
\value{
a list included the correlation, \eqn{r}, the sample size, \eqn{n},
the t-test statistic, \eqn{t}, the p-value, \eqn{p}, and the standard error, \eqn{se}
}
\description{
Compute the corresponding p-value for a given correlation, \eqn{r}, and sample size, \eqn{n}.
}
