% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Algorithm1.R
\name{create_structure_list}
\alias{create_structure_list}
\title{Creates a list of candidate structures for the SLIDE model}
\usage{
create_structure_list(X, pvec, lambda_seq = NULL, n_lambda = 50,
  lambda_min = 0.01, lambda_max = 1, ratio_max = 2/3, standardized = F,
  eps = 1e-06, k_max = 1000)
}
\arguments{
\item{X}{A n x p concatenated data matrix of views X_1,...,X_d.}

\item{pvec}{A vector of values p_1,....,p_d corresponding to the number of measurements within each  view.}

\item{lambda_seq}{An optional sequence of tuning parameters for the penalized matrix decomposition problem. By default, the algorithm generates its own sequence based on supplied values of \code{n_lambda}, \code{lambda_min} and \code{lambda_max}.}

\item{n_lambda}{A length of tuning parameter sequence. The default value is 50. It is only used when \code{lambda_seq = NULL}.}

\item{lambda_min}{A minimal value for tuning parameter. The default value is 0.01.}

\item{lambda_max}{A maximal value for tuning parameter. The default value is 1. If X is already standardized, it is recommended to set \code{lambda_max} to the largest singular value within the view. If \code{standardized = F}, this choice is made automatically by the algorithm.}

\item{ratio_max}{A maximal allowable rank of the binary structure as a ratio of maximal rank of X, the default value is 2/3.}

\item{standardized}{A logical indicator of whether X is centered and standardized. The default value is FALSE and the standardization is performed within the function.}

\item{eps}{A convergence tolerance criterion as measured by the differene in objective functions at successive iterations, the default value is 1e-06.}

\item{k_max}{A maximal number of allowed iterations, the default value is 1000.}
}
\value{
A list with the elements
  \item{lambda_seq}{A sequence of tuning parameters used to generate the structure sequence.}
  \item{Slist}{A list of distinct binary structures which forms a candidate set for SLIDE model.}
  \item{id_used}{A binary vector indicating the correspondence between \code{lambda_seq} and \code{Slist}. The zero values indicate the tuning parameters that either lead to the same structures, or resulted in structures with the rank larger than allowed by \code{ratio_max}.}
}
\description{
Creates a list of candidate structures for the SLIDE model based on the solution path of penalized matrix factorization problem.
}
\details{
The function solves the penalized matrix factorization problem for each value of from the sequence of tuning parameters lambda_1,..., lambda_m. The block-sparsity pattern of the resulting V(lambda_1), ..., V(lambda_m) is used to generate the sequence of binary structures S(lambda_1),..., S(lambda_m). This sequence is further trimmed to remove any repetitions, and to only keep structures with the number of nonzero columns being less than \code{ratio_max * min(n, sum(pvec))}.
}
\examples{
n = 50
p1 = 20
p2 = 40
X1 = matrix(rnorm(n*p1), n, p1)
X2 = matrix(rnorm(n*p2), n, p2)
X = cbind(X1, X2)
pvec = c(p1,p2)

# Unstandardized
slist = create_structure_list(X, pvec)

# Already standardized
out = standardizeX(X, pvec = c(p1,p2))
slist = create_structure_list(out$X, pvec, lambda_max = max(out$svec), standardized = TRUE)
}
