% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cvDelta.R
\name{cvDelta}
\alias{cvDelta}
\title{\eqn{\delta} Cross-Validation.}
\usage{
cvDelta(raw_x, fdr_entries, deltas_scaled)
}
\arguments{
\item{raw_x}{data matrix of dimensions \eqn{n \times p}, not standardized}

\item{fdr_entries}{a matrix of dimensions \eqn{p \times p} that contains a 1 in positions where the
entry is kept in the thresholded version of \eqn{\hat{\Sigma}} and a 0 if not}

\item{deltas_scaled}{a vector of numerical constants over which to perform the search for the optimal \eqn{\delta}}
}
\value{
the selected optimal \eqn{\delta}
}
\description{
Cross validation for choosing \eqn{\delta}. For each delta from the given grids,
first split the data into two data sets. Obtain \eqn{I}, \eqn{A_I} and \eqn{C} from data set 1.
Then calculate \eqn{A_I \cdot C \cdot A_I^\top} and choose \eqn{\delta} which minimizes the criterion
\eqn{A_I \cdot C \cdot A_I^\top - \Sigma(\text{data set 2})}.
}
