% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cvLambda.R
\name{cvLambda}
\alias{cvLambda}
\title{\eqn{\lambda} Cross-Validation.}
\usage{
cvLambda(x, fdr_entries, lambdas, AI, pure_vec, k)
}
\arguments{
\item{x}{a data matrix of dimensions \eqn{n \times p}}

\item{fdr_entries}{a matrix of dimensions \eqn{p \times p} that contains a 1 in positions where the
entry is kept in the thresholded version of \eqn{\hat{\Sigma}} and a 0 if not}

\item{lambdas}{a vector of numerical constants over which to search for the optimal \eqn{\lambda}}

\item{AI}{the estimated matrix \eqn{A_I} of dimensions \eqn{p \times K}}

\item{pure_vec}{a vector of indices of pure variables}

\item{k}{number of folds for \eqn{k}-fold cross-validation}
}
\value{
the selected optimal \eqn{\lambda}
}
\description{
Use cross-validation to select \eqn{\lambda} for estimating \eqn{\Omega}.
Split the data into two parts and estimate \eqn{C} on both data sets. Then, for each
\eqn{lambda}, calculate \eqn{\Omega} on the first dataset and calculate the loss on the second dataset.
Find the lambda which minimizes \eqn{C \cdot \Omega - log(|\Omega|)}.
}
