% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dantzig.R
\name{dantzig}
\alias{dantzig}
\title{Dantzig Estimation.}
\usage{
dantzig(C_hat, target_vec, lambda)
}
\arguments{
\item{C_hat}{a matrix}

\item{target_vec}{a vector}

\item{lambda}{\eqn{\lambda}, a positive constant}
}
\value{
a row of \eqn{A_J} estimated with a Dantzig-type estimator
}
\description{
Estimate a vector using a Dantzig-type estimator.
}
