% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/essregCV.R
\name{essregCV}
\alias{essregCV}
\title{Cross-Validation For Essential Regression.}
\usage{
essregCV(
  k = 5,
  y,
  x,
  delta,
  std_cv,
  std_y,
  thresh_fdr = 0.2,
  lambda = 0.1,
  rep_cv = 50,
  alpha_level = 0.05,
  permute = T,
  y_levels = NULL,
  eval_type,
  out_path,
  rep,
  benchmark = F,
  run_lasso = F
)
}
\arguments{
\item{k}{an integer for number of folds to use in cross-validation}

\item{y}{a response vector of dimension \eqn{n}}

\item{x}{a data matrix of dimensions \eqn{n \times p}}

\item{delta}{\eqn{\delta}, a numerical constant used for thresholding}

\item{std_cv, }{a boolean flag of whether perform scaling x in CV and scaling Y}

\item{std_y}{a boolean flag of wether z score Y or not in the PlainER function}

\item{thresh_fdr}{a numerical constant used for thresholding the correlation matrix to
control the false discovery rate, default is 0.2}

\item{lambda}{\eqn{\lambda}, a numerical constant used in thresholding}

\item{rep_cv}{number of replicates for cross-validation}

\item{alpha_level}{\eqn{\alpha}, a numerical constant used in confidence interval calculation}

\item{permute}{a boolean flag indicating whether to permute the response (\eqn{y}) during training}

\item{eval_type}{a string indicating what metric to use for evaluating model performance
(can be "auc" - area under ROC curve, or "corr" - for spearman correlation)}

\item{out_path}{path for saving output}

\item{rep}{an integer indicating the replicate number}

\item{benchmark}{a boolean flag that decides if running lasso, pcr, pclr, plsr and plsda.}

\item{run_lasso}{a boolean flag that decides if running lasso for comparison}
}
\value{
An object of class \sQuote{data.frame}
}
\description{
Perform k-fold cross-validation for essential regression to select.
}
