% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estAI.R
\name{estAI}
\alias{estAI}
\title{Estimate \eqn{A_I}.}
\usage{
estAI(sigma, delta, se_est)
}
\arguments{
\item{sigma}{a correlation matrix of dimensions \eqn{p \times p}}

\item{delta}{\eqn{\delta}, a numerical constant}

\item{se_est}{standard deviations of the columns of the data matrix, \code{x}}
}
\value{
a list including: \eqn{A_I}, a matrix of dimensions \eqn{p \times K},
a vector of the indices of estimated pure variables, and a list of the indices of
the estimated pure variables
}
\description{
Use the given \eqn{\delta} to calculate the fitted \eqn{A_I}, pure variable
indices in list form and vector form. Also return estimated \eqn{Y} and \eqn{C} for
the following Dantzig estimation.
}
