% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estAJDant.R
\name{estAJDant}
\alias{estAJDant}
\title{Estimation \eqn{A_J} (Dantzig).}
\usage{
estAJDant(C_hat, sigma_TJ, lambda, se_est_J)
}
\arguments{
\item{C_hat}{a matrix}

\item{sigma_TJ}{a matrix}

\item{lambda}{\eqn{\lambda}, a positive constant}

\item{se_est_J}{the estimated standard errors of indices in \eqn{J}}
}
\value{
an estimate of matrix \eqn{A_J}
}
\description{
Estimate the matrix \eqn{A_J} with a Dantzig-type estimator.
}
