% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estBeta.R
\name{estBeta}
\alias{estBeta}
\title{Estimate \eqn{\beta}.}
\usage{
estBeta(
  y,
  x,
  sigma,
  A_hat,
  C_hat,
  Gamma_hat,
  I_hat,
  I_hat_list,
  alpha_level = 0.05
)
}
\arguments{
\item{y}{vector of responses of dimension \eqn{p}}

\item{x}{data matrix of dimensions \eqn{n \times p}}

\item{sigma}{sample correlation matrix of the data, \eqn{X}}

\item{A_hat}{estimated matrix of dimensions \eqn{p \times K}}

\item{C_hat}{estimated matrix}

\item{Gamma_hat}{estimated matrix}

\item{I_hat}{estimated matrix}

\item{I_hat_list}{estimated matrix}

\item{alpha_level}{value for confidence intervals}
}
\value{
a list including the estimates for \eqn{\beta} and the confidence intervals (if requested)
}
\description{
Estimate the values of the coefficients \eqn{\beta} as well as confidence intervals for the estimates
if requested.
}
