% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estOmega.R
\name{estOmega}
\alias{estOmega}
\title{Estimate \eqn{\Omega}.}
\usage{
estOmega(lambda, C)
}
\arguments{
\item{lambda}{\eqn{\lambda}}

\item{C}{a square, symmetric matrix}
}
\value{
\eqn{\Omega = C^{-1}}
}
\description{
For a given \eqn{\lambda} and \eqn{C}, find \eqn{C^{-1}}.
}
