% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estSigmaAlt.R
\name{estSigmaAlt}
\alias{estSigmaAlt}
\title{estSigmaAlt.}
\usage{
estSigmaAlt(y, h_hat, beta_hat, C_hat)
}
\arguments{
\item{y}{a response vector of dimension \eqn{p}}

\item{h_hat}{estimated matrix:
\eqn{\frac{1}{n}(\hat{A}_{\hat{I}}^\top \hat{A_{\hat{I}}^\top)^{-1}\hat{A}_{\hat{I}}^\top \mathbf{X}_{\hat{I}}^\top \mathbf{Y}})}}

\item{beta_hat}{estimated values of \eqn{\beta}}

\item{C_hat}{estimated matrix}
}
\value{
???
}
\description{
I am not sure what this does.
}
