% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estSigmaTJ.R
\name{estSigmaTJ}
\alias{estSigmaTJ}
\title{Estimate \eqn{\Sigma_{\hat{TJ}}}.}
\usage{
estSigmaTJ(sigma, AI, pure_vec)
}
\arguments{
\item{sigma}{a sample correlation matrix of dimensions \eqn{p \times p}}

\item{AI}{a matrix of dimensions \eqn{p \times K}}

\item{pure_vec}{a vector of indices of the pure nodes}
}
\value{
an estimate for \eqn{\Sigma_{\hat{TJ}}} of dimensions \eqn{K \times |J|}
}
\description{
Estimate the matrix, \eqn{\Sigma_{\hat{TJ}}}.
}
