% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findFeats.R
\name{findFeats}
\alias{findFeats}
\title{Find the features of interest in a given latent factor.}
\usage{
findFeats(x, y, z, er_res, thresh = 0.01, p_thresh = 0.1)
}
\arguments{
\item{x}{a matrix or data frame of numeric values; the features}

\item{y}{a vector or data frame with one column of numeric values; the true response}

\item{z}{an integer; the column index of the latent factor of interest}

\item{er_res}{a list; the results from running Essential Regression}

\item{thresh}{a numeric value; the threshold to use to determine significance for the absolute loadings of each feature in \eqn{A}}

\item{p_thresh}{a numeric value; the threshold to use to determine significance for the univariate p-values}
}
\value{
a list including the pure variables, mixed variables, features with significant p-values, features with
significantly large loadings, correlations between response and features, and the univariate p-values for the
latent factor of interest
}
\description{
For a given latent factor of interest, extract the features with the highest loadings in the
\eqn{A} matrix calculated by Essential Regression, find the Spearman correlation between the response (\eqn{y})
and these features, and find the p-values on the coefficients of these features from
a simple linear regression of \eqn{y} on these features (univariate).
}
