% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findFrequent.R
\name{findFrequent}
\alias{findFrequent}
\title{Find most frequent variables.}
\usage{
findFrequent(spec, niter, tab_data)
}
\arguments{
\item{spec}{a numeric constant between 0 and 1; the change in frequency proportion}

\item{niter}{an integer; the number of times knockoffs was run}

\item{tab_data}{the frequency table made after running iterated knockoffs}
}
\value{
a vector of names corresponding to the selected columns of \code{z} that are determined to be frequent enough
}
\description{
Finds the "joint" of the elbow if a histogram were to be made. Loops through all variables and selects those that appear in at least
50\% of the iterations or that are before the elbow. In this case, the elbow occurs when the proportion of selections in the iterations
changes by more than \code{spec}.
}
