% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findOptIter.R
\name{findOptIter}
\alias{findOptIter}
\title{Find the best results from \code{niter} iterations of knockoffs.}
\usage{
findOptIter(freq_vars, selected_list)
}
\arguments{
\item{freq_vars}{a vector; the names of the most frequent variables}

\item{selected_list}{a list of the selected variables each iteration of knockoffs}

\item{n}{an integer; sample size}
}
\value{
a vector containing the names of the variables found in the results of the best iteration
}
\description{
Find the variables that are selected at least \code{spec} proportion of the \code{niter} of the iterations then find the shortest results from all \code{niter} that
have the maximum overlap with this list of selected variables.
}
