% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findPureNode.R
\name{findPureNode}
\alias{findPureNode}
\title{Estimate Pure Nodes.}
\usage{
findPureNode(abs_sigma, delta, max_vals, max_inds, se_est)
}
\arguments{
\item{abs_sigma}{a sample correlation matrix of dimensions \eqn{p \times p} with entries
in absolute value and diagonal of 0}

\item{delta}{\eqn{\delta}, a numerical constant}

\item{max_vals}{the largest absolute values of each row of \code{abs_sigma}}

\item{max_inds}{vector of column indices at which the values in \code{max_vals} are achieved in \code{abs_sigma}}

\item{se_est}{standard deviations of features (columns of \code{x})}
}
\value{
A list including the list of estimated pure node indices and a vector
of the estimated pure node indices.
}
\description{
Estimate list of pure node indices for given \eqn{\Sigma} and \eqn{\delta}.
This code is an implementation of Algorithm 1 from Bing et al. (2020).
}
