% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findRowMax.R
\name{findRowMax}
\alias{findRowMax}
\title{Find \eqn{\Sigma} Maxima.}
\usage{
findRowMax(abs_sigma)
}
\arguments{
\item{abs_sigma}{a matrix of dimensions \eqn{p \times p} with entries in absolute value and
diagonal of 0s}
}
\value{
a list including the index of the (first) maximal absolute values of the rows of
\eqn{\Sigma} as well as the values
}
\description{
Calculate the maximal absolute value for each row of the given matrix.
}
