% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findSignPureNode.R
\name{findSignPureNode}
\alias{findSignPureNode}
\title{Estimate Signed Subpartitions Of Pure Nodes.}
\usage{
findSignPureNode(pure_list, sigma)
}
\arguments{
\item{pure_list}{a list of pure node indices}

\item{sigma}{a sample correlation matrix of dimensions \eqn{p \times p}}
}
\value{
a list of signed subpartitions of pure node indices
}
\description{
Estimate the signed subpartition of pure node sets. If there is an element
of a list is empty, then a empty list will be put in that position.
}
