% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Models.R
\name{generateModel1}
\alias{generateModel1}
\title{Generate data from SLIDE model for two views.}
\usage{
generateModel1(
  n = 100,
  pvec = c(25, 25),
  cvec = c(1, 1),
  snr = 1,
  orthogonalV = T
)
}
\arguments{
\item{n}{An integer specifying the number of matched samples, the default value is 100.}

\item{pvec}{A vector of integer values p_1, p_2 corresponding to the number of measurements within each data view. The default value is \code{pvec = c(25,25)}.}

\item{cvec}{A vector of positive values c_1, c_2 corresponding to the scale of each view. The default value is \code{cvec = c(1,1)}, that is both views have the same scale.}

\item{snr}{A non-negative value specifying the ratio between the Frobenius norm of the signal matrix and the Frobenius norm of the noise matrix. This value determines the variance of the elements in the noise matrix. The default value is 1.}

\item{orthogonalV}{A logical indicating whether the loading vectors should be generated orthogonal to each other, the default value is TRUE.}
}
\value{
A list with the elements
  \item{X}{A n x (p_1 + p_2) concatenated data matrix of views X_1, X_2.}
  \item{pvec}{The supplied values of \code{pvec}.}
  \item{sigma1}{The standard deviation of the elements of the noise matrix for the 1st dataset.}
  \item{sigma2}{The standard deviation of the elements of the noise matrix for the 2nd dataset.}
  \item{U}{The score matrix from the SLIDE model.}
  \item{V}{The loadings matrix from the SLIDE model.}
  \item{cvec}{The supplied values of \code{cvec}.}
  \item{snr}{The supplied value of \code{snr}.}
}
\description{
Generates two matched datasets from SLIDE model with rank 2 for each type of structure (shared/individual).
}
\details{
Generates X_1 (n x p_1) and X_2 (n x p_2) according to the additive noise model \deqn{X_1 = Z_1 + E_1, X_2 = Z_2 + E_2,} where noise matrices \eqn{E_1, E_2} have independent entries from normal distribution with mean zero, and standard deviations sigma1, sigma2 correspondingly, and
\deqn{Z_1 = c_1 (U_1 D_0 W_{1,1}' + U_2 D_1 W_{1,2}')}
\deqn{Z_2 = c_2 (U_1 D_0 W_{2,1}' + U_3 D_2 W_{2,3}').}
Here \eqn{U=[U_1 U_2 U_3]} (n x 6) is generated using uniform distribution on [0,1] with subsequent centering and orthonormalization, \eqn{D_0 = diag(1.5, 1.3)}, \eqn{D_1 = diag(1, 0.8)}, \eqn{D_2 = diag(1, 0.7)}, and the loadings \eqn{W = [W_1 W_2]} (p_1 + p_2 X 6) with \eqn{W_1 = [W_{1,1} W_{1,2} 0]}, \eqn{W_2 = [W_{2,1} 0 W_{2,3}]} are generated using uniform distribution on [0,1] with subsequent orthonormalization of \eqn{W}. The loadings matrix \eqn{V} in SLIDE model is generated by condensing c_1, c_2, D_0, D_1, D_2 and W.
}
\examples{
data <- generateModel1(n = 50, pvec = c(50,75), snr = 2)
data <- generateModel1(n = 100, pvec = c(25,150), snr = 1)
data <- generateModel1(n = 100, pvec = c(25,25),cvec = c(0.5,1.5), snr = 1)
}
