% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFolds.R
\name{getFolds}
\alias{getFolds}
\title{Split data for cross-validation.}
\usage{
getFolds(k, x, y)
}
\arguments{
\item{k}{an integer; the number of folds}

\item{x}{a data frame or matrix; the data}

\item{y}{a data frame or vector; the reponse}
}
\value{
a list of vectors containing the indices of the rows of the data found in each fold
}
\description{
Randomly split data into \code{k} folds for cross-validation.
}
