% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getLatentFactors.R
\name{getLatentFactors}
\alias{getLatentFactors}
\title{Get Latent Factors from Input Data.}
\usage{
getLatentFactors(
  y,
  x,
  x_std,
  std_y = TRUE,
  sigma = NULL,
  delta,
  thresh_fdr = 0.2,
  lambda = 0.1,
  rep_cv = 50,
  alpha_level = 0.05,
  out_path = NULL,
  save_heatmaps = FALSE
)
}
\arguments{
\item{y}{a response vector of dimension \eqn{n}, must be continous}

\item{x}{a data matrix of dimensions \eqn{n \times p}}

\item{x_std}{a data matrix of dimensions \eqn{n \times p}, a scaled and mean centered x matrix}

\item{std_y}{a boolean flag of wether z score Y or not}

\item{sigma}{a sample correlation matrix of dimensions \eqn{p \times p}, default is \code{NULL}}

\item{delta}{\eqn{\delta}, a numerical value used for thresholding, or a vector of values}

\item{thresh_fdr}{a numerical constant used for thresholding the correlation matrix to
control the false discovery rate, default is 0.2}

\item{lambda}{\eqn{\lambda}, a numerical constant used in thresholding}

\item{rep_cv}{number of replicates for cross-validation}

\item{alpha_level}{\eqn{\alpha}, a numerical constant used in confidence interval calculation}

\item{out_path}{a string path to where to save output}

\item{save_heatmaps}{a logical flag indicating whether to save correlation matrix heatmaps}
}
\value{
a list of results from the Essential Regression framework including: \eqn{K} = number of clusters,
\eqn{\hat{A}}, \eqn{\hat{C}}, \eqn{\hat{I}}, the indices of the pure variables, \eqn{\hat{\Gamma}},
\eqn{\hat{\beta}}, \eqn{\alpha}-level confidence intervals (if requested), prediction results (if requested),
the optimal value of \eqn{\lambda} determined by cross-validation, the optimal value of \eqn{\delta}
determined by cross-validation, \eqn{Q}, and the variances of \eqn{\hat{\beta}}
}
\description{
Perform Essential Regression without prior knowledge.
}
