% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTopBetas.R
\name{getTopBetas}
\alias{getTopBetas}
\title{Method A: Top Betas.}
\usage{
getTopBetas(betas, top_prop)
}
\arguments{
\item{betas}{a vector of values for the \eqn{\beta}s}

\item{top_prop}{a numeric constant: the proportion of \eqn{\beta}s to consider significant}
}
\value{
a vector of \eqn{Z} indices
}
\description{
Using the provided proportion, retrieve the top \eqn{\beta}s by absolute size and identify the corresponding latent \eqn{Z}s.
}
