% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getUpsilon.R
\name{getUpsilon}
\alias{getUpsilon}
\title{Find value of upsilon.}
\usage{
getUpsilon(interaction_var, interactions, z, y, marg_var)
}
\arguments{
\item{interaction_var}{a vector; the names of the significant interaction terms}

\item{interactions}{a matrix or data frame; the values of the significant interaction terms}

\item{z}{a matrix or data frame}

\item{marg_var}{an integer; the index of a marginal variable}
}
\value{
a data frame with 1 column
}
\description{
Find upsilon by extracting the fitted values from a linear regression model.
}
