% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slide_data.R
\docType{data}
\name{I_sig}
\alias{I_sig}
\title{Protein Expression Levels in an Infected Cell Sample}
\format{An object of class \code{data.frame} with 38 rows and 7 columns.}
\usage{
I_sig
}
\description{
Simulated data from an experiment measuring viral reprogramming of lymphocyte surface proteins. 
These cells demonstrate evidence for significant reprogramming (see vignette for example).
}
\keyword{datasets}
