% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indName.R
\name{indName}
\alias{indName}
\title{Indices To Names.}
\usage{
indName(feats, all_names, to_ind = T)
}
\arguments{
\item{feats}{a vector of feature indices or names}

\item{all_names}{a vector of all feature names}

\item{to_ind}{a boolean indicating whether to go from indices to names (T) or names to indices (F)}
}
\value{
a vector of feature indices or names depending upon value of \code{toInd}
}
\description{
Translate feature indices to feature names and vice versa.
}
