% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactionSLIDE.R
\name{interactionSLIDE}
\alias{interactionSLIDE}
\title{Find significant interaction terms.}
\usage{
interactionSLIDE(
  z,
  y,
  m,
  marginals,
  method = 4,
  elbow = FALSE,
  spec = 0.3,
  fdr = 0.1,
  niter = 1000,
  f_size = 100,
  parallel = TRUE
)
}
\arguments{
\item{z}{a matrix or data frame of measured data of dimensions \eqn{n} by \eqn{p}}

\item{y}{a matrix or data frame of responses of dimensions \eqn{n} by 1}

\item{m}{an integer; the target model size}

\item{method}{an integer; the selected method to use}

\item{elbow}{a boolean flag; whether to select using \code{spec} or by identifying the "joint" of the histogram elbow}

\item{spec}{a numeric constant between 0 and 1; the proportion of iterations that a variable must be selected in to be considered significant}

\item{fdr}{a numeric constant between 0 and 1; the target false discovery rate for knockoffs}

\item{niter}{an integer; the number of times to run knockoffs}

\item{f_size}{an integer; the target size for the subset (number of columns in each subset)}

\item{parallel}{a boolean flag; run iterations in parallel or sequentially}
}
\value{
a list containing the names of the significant interaction terms, the values of those terms, and the values of \eqn{\upsilon}
(NULL if not method 4)
}
\description{
Iterate through marginal variables and construct and test their interaction terms.
Concatenate all marginal variable results.
}
