% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interUnion.R
\name{interUnion}
\alias{interUnion}
\title{Construct interaction terms according to union logic.}
\usage{
interUnion(marginal_vars, z)
}
\arguments{
\item{marginal_vars}{a vector of significant marginal variable indices}

\item{z}{a matrix or data frame; the values for all variables}
}
\value{
a list including the significant marginal variables that were used as a vector and a matrix
containing the values of the constructed interaction terms
}
\description{
Given a list of significant marginal variables, construct their interaction terms with all other
variables without repetitions (i.e. if \eqn{Z1.Z2} is already made, do not create \eqn{Z2.Z1}).
}
