% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LP.R
\name{LP}
\alias{LP}
\title{Linear Program Function.}
\usage{
LP(y, lambda)
}
\arguments{
\item{y}{response vector of dimension \eqn{K}}

\item{lambda}{\eqn{\lambda}, a positive constant}
}
\value{
a vector of dimension \eqn{K}
}
\description{
Solve the following using linear programming:
\deqn{\beta^{+} - \beta^{-} \leq \lambda + y - \beta^{+} + \beta^{-} \leq \lambda - y}
for semipositive \eqn{\beta^{+}, \beta^{-}}
}
