% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{main}
\alias{main}
\title{Main Pipeline of Running SLIDE (get LF and run SLIDE without CV)}
\usage{
main(yaml_path, sink_file)
}
\arguments{
\item{yaml_path}{the string to a yaml path}

\item{sink_file}{boolean flag of saving to a sink file or not}
}
\description{
Main Pipeline of Running SLIDE (get LF and run SLIDE without CV)
}
