% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeHeatmap.R
\name{makeHeatmap}
\alias{makeHeatmap}
\title{Create Matrix Heatmap.}
\usage{
makeHeatmap(mat, title, cluster = T, names = F)
}
\arguments{
\item{mat}{the matrix for plotting}

\item{title}{a string for the plot title}

\item{cluster}{a boolean indicating whether to perform hierarchical clustering}

\item{names}{a boolean indicating whether to display column/row names}
}
\value{
a heatmap
}
\description{
Create a heatmap for the provided matrix. This is essentially just a wrapper for
\code{pheatmap}.
}
