% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marginalSLIDE.R
\name{marginalSLIDE}
\alias{marginalSLIDE}
\title{Find marginal variables.}
\usage{
marginalSLIDE(
  z,
  y,
  method = 4,
  betas = NULL,
  top_prop = NULL,
  marginals = NULL,
  elbow = FALSE,
  spec = 0.5,
  fdr = 0.1,
  niter = 1000,
  f_size = 100,
  parallel = TRUE
)
}
\arguments{
\item{z}{a matrix or data frame of measured data of dimensions \eqn{n} by \eqn{p}}

\item{y}{a matrix or data frame of responses of dimensions \eqn{n} by 1}

\item{method}{an integer; the selected method to use}

\item{betas}{a vector of numeric values; the coefficients corresponding to the variables in \code{z} (found using Essential Regression)}

\item{top_prop}{a numeric constant between 0 and 1; the proportion of \code{betas} to consider significant}

\item{marginals}{a vector of numeric values; the indices of the variables to consider significant}

\item{elbow}{a boolean flag; whether to select using \code{spec} or by identifying the "joint" of the histogram elbow}

\item{spec}{a numeric constant between 0 and 1; the proportion of iterations that a variable must be selected in to be considered significant}

\item{fdr}{a numeric constant between 0 and 1; the target false discovery rate for knockoffs}

\item{niter}{an integer; the number of times to run knockoffs}

\item{f_size}{an integer; the target size for the subset (number of columns in each subset)}

\item{parallel}{a boolean flag; run iterations in parallel or sequentially}
}
\value{
a vector of names corresponding to the selected columns of \code{z} that are determined to be significant
}
\description{
Find significant marginal variables. There are currently six supported methods:
top_betas - Significant marginal variables are selected as the variables corresponding to the absolute largest coefficients from Essential Regression.
Selected variables are in the top \eqn{\frac{\text{spec}}{2}} proportion of the positive or negative largest coefficients.
marginals - Significant marginal variables are selected as the provided vector of indices. This requires passing in a vector of indices as \code{marginals}.
1 - Significant marginal variables are those appearing in the longest list returned from \code{niter} iterations of
Gaussian knockoffs
2 - Significant marginal variable is the single most freqently appearing variable returned from \code{niter} iterations of
second order knockoffs
3 - Significant marginal variables are those appearing in the best list returned from \code{niter} iterations of
second order knockoffs. In this case, "best" refers to the shortest list that has the maximal overlap with the list of
most frequently selected variables. The frequency list is comprised of those variables appearing at least \code{spec}
proportion of the interactions. If the number of selected significant marginal variables is longer than the number of samples
(number of rows of \code{z}), then \code{spec} is increased by 0.1, and the selected process is repeated.
4 - Significant marginal variables are selected in the same way as in Method 3 but without reselection in the case of too long a
list.
}
