% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge.R
\name{merge}
\alias{merge}
\title{Merge Nodes (Intersection). NOT USED ANYMORE}
\usage{
merge(clusters, group)
}
\arguments{
\item{clusters}{a list of groups of node indices}

\item{group}{a vector of node indices}
}
\value{
a list of the merged results
}
\description{
Merge a group with other groups containing common nodes.
}
