% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergeUnion.R
\name{mergeUnion}
\alias{mergeUnion}
\title{Merge Nodes (Union).}
\usage{
mergeUnion(clusters, group)
}
\arguments{
\item{clusters}{a list of groups of node indices}

\item{group}{a vector of node indices}
}
\value{
a list of the merged results
}
\description{
Merge the provided vector of nodes into the list of vectors of nodes but use
union rather than the intersection.
}
