% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/offSum.R
\name{offSum}
\alias{offSum}
\title{Off-Diagonal Sum Of Squares.}
\usage{
offSum(mat1, mat2, weights)
}
\arguments{
\item{mat1}{a matrix}

\item{mat2}{a matrix of same dimensions as \code{mat1}}

\item{weights}{the weights to use in calculating the sum}
}
\value{
the sum of squares
}
\description{
Calculate the sum of squares of the upper off-diagonal elements of two matrices.
}
