% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Optimize_SLIDE.R, R/main.R
\name{Optimize_SLIDE}
\alias{Optimize_SLIDE}
\title{Main Pipeline of Running SLIDE (get LF and run SLIDE without CV)}
\usage{
Optimize_SLIDE(input_params, sink_file)

Optimize_SLIDE(input_params, sink_file)
}
\arguments{
\item{input_params}{the R object that contains all the input parameter}

\item{sink_file}{boolean flag of saving to a sink file or not}

\item{yaml_path}{the string to a yaml path}
}
\description{
Main Pipeline of Running SLIDE (get LF and run SLIDE without CV)

Main Pipeline of Running SLIDE (get LF and run SLIDE without CV)
}
