% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimizeSLIDE.R
\name{optimizeSLIDE}
\alias{optimizeSLIDE}
\title{Main Pipeline of Running SLIDE (get LF and run SLIDE without CV)}
\usage{
optimizeSLIDE(input_params, sink_file = F)
}
\arguments{
\item{sink_file}{boolean flag of saving to a sink file or not}

\item{yaml_path}{the string to a yaml path}
}
\description{
Main Pipeline of Running SLIDE (get LF and run SLIDE without CV)
}
