% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwiseInteractions.R
\name{pairwiseInteractions}
\alias{pairwiseInteractions}
\title{PairwiseInteractions}
\usage{
pairwiseInteractions(index_list, mat)
}
\arguments{
\item{index_list}{a string that points to the x matrix.}

\item{mat}{a string that points to the final er result as an RDS object.}
}
\value{
a matrix of pairwise interactions
}
\description{
PairwiseInteractions
}
