% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paperBench.R
\name{paperBench}
\alias{paperBench}
\title{Benchmarking methods with cross-validation wrapper.}
\usage{
paperBench(yaml_path, replicate)
}
\arguments{
\item{yaml_path}{a string path to the .yaml file with run configuration}

\item{replicate}{an integer; the replicate number used for output directory naming}
}
\value{
none
}
\description{
A wrapper for \code{benchCV} that configures the run by reading in a provided
.yaml file and creates the results directories.
}
