% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseRun.R
\name{parseRun}
\alias{parseRun}
\title{Parse .yaml file for Essential Regression}
\usage{
parseRun(yaml_path)
}
\arguments{
\item{yaml_path}{a string path to the .yaml file}
}
\value{
none
}
\description{
Read a .yaml file, parse arguments, and run Essential Regression.
}
