% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permA.R
\name{permA}
\alias{permA}
\title{Find Optimal Permuted Matrix.}
\usage{
permA(A, B)
}
\arguments{
\item{A}{a matrix}

\item{B}{a matrix of same dimensions as \eqn{A}}
}
\value{
the optimal permutation of \eqn{A}
}
\description{
Find \eqn{A_{perm}}, the optimal permuted matrix of \eqn{A}, such that
\eqn{|A_{perm} - B|_F} is minimized.
}
