% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipelineER1.R
\name{pipelineER1}
\alias{pipelineER1}
\title{Essential Regression Pipeline - Steps 1, 2}
\usage{
pipelineER1(yaml_path, steps = "all")
}
\arguments{
\item{yaml_path}{the path to a .yaml file containing all necessary parameters/arguments
for Essential Regression}

\item{steps}{an integer or string indicating which steps of the pipeline to perform: "1" or "all"}
}
\value{
nothing is returned, saves boxplot of cross-validation results for user to use
in selecting optimal \eqn{\delta} if \code{steps} is "all"
}
\description{
Run Essential Regression Pipeline with K-Fold Cross-Validation
Step 1 - Coarse Grid Search for \eqn{\delta}
Step 2 - K-Fold Cross-Validation to find magnitude of \eqn{\delta}
}
