% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipelineER3.R
\name{pipelineER3}
\alias{pipelineER3}
\title{Essential Regression Pipeline - Step 5}
\usage{
pipelineER3(yaml_path)
}
\arguments{
\item{yaml_path}{the path to a .yaml file containing all necessary parameters/arguments
for Essential Regression}
}
\value{
nothing is returned, saves boxplot of cross-validation results and final ER output
}
\description{
Run Essential Regression Pipeline with K-Fold Cross-Validation
Step 5 - Run final replicates of K-Fold CV using optimal \eqn{\delta} and optimal \eqn{\lambda}.
Create a boxplot and run \code{plainER()} one last time for final results.
}
