% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSigGenes.R
\name{plotSigGenes}
\alias{plotSigGenes}
\title{Plot significant genes for significant (marginal) latent factors}
\usage{
plotSigGenes(slide_results, plot_interactions = F, out_path = NULL)
}
\arguments{
\item{slide_results}{list - SLIDE results output from runSLIDE function}

\item{plot_interactions}{logical - whether to plot interaction variables}

\item{out_path}{string - Output path to save results}
}
\description{
Plot significant genes for significant (marginal) latent factors
}
