% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotViol.R
\name{plotViol}
\alias{plotViol}
\title{Create a violin plot of results.}
\usage{
plotViol(yaml_path)
}
\arguments{
\item{yaml_path}{a string path to the .yaml file that configured the cross-validation}
}
\value{
none
}
\description{
Reads in results from running the benchmarking cross-validation regime and creates a
violin plot of the replicate results.
}
