% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predError.R
\name{predError}
\alias{predError}
\title{Calculate prediction error.}
\usage{
predError(
  x,
  y,
  spec = 0.3,
  niter = 1000,
  fdr = 0.1,
  f_size = 100,
  ncore = 64,
  out_file,
  k = NULL
)
}
\arguments{
\item{x}{a data frame of numeric values; the data}

\item{y}{a vector or data frame with one column; the response}

\item{spec}{a numeric constant; the proportion of iterations that a variable must be selected in to be considered significant}

\item{niter}{an integer; the number of times to run knockoffs}

\item{fdr}{a numeric constant between 0 and 1; the target false discovery rate for knockoffs}

\item{f_size}{an integer; the target size for the subset (number of columns in each subset)}

\item{ncore}{an integer; the number of cores to use for parallel computing}

\item{out_file}{a string path/file name; where to save the results}
}
\value{
a data frame with 5 rows and 2 columns (method name and method MSE)
}
\description{
The meat of \code{runSims()}. This function splits the data into a training set
and a validation set, trains the methods, and then calculates the error (MSE) using the
validation set.
}
