% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recoverAI.R
\name{recoverAI}
\alias{recoverAI}
\title{Find \eqn{A_I} From Pure Nodes.}
\usage{
recoverAI(pure_list, p)
}
\arguments{
\item{pure_list}{a list of group indices of the pure nodes with sign}

\item{p}{the number of features}
}
\value{
a matrix of dimensions \eqn{p \times K}
}
\description{
Recover the estimated submatrix \eqn{A_I} by given the pure node group.
}
