% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runLASSO.R
\name{runLASSO}
\alias{runLASSO}
\title{Run LASSO regression for cross-validation.}
\usage{
runLASSO(k, train_y, train_x, valid_x, lasso_fam, y_factor)
}
\arguments{
\item{k}{the number of folds for cross-validation}

\item{train_y}{a vector of numeric values; the response (training set)}

\item{train_x}{a matrix or data frame of numeric values; the data (training set)}

\item{valid_x}{a matrix or data frame of numeric values; the data (validation set)}

\item{lasso_fam}{a string indicating whether to perform linear or logistic regression}

\item{y_factor}{a boolean flag; whether the response is categorical or not}
}
\value{
a list containing the training model and the predicted values of the response
using \code{valid_x} and \code{train_x}
}
\description{
Runs LASSO to create a regression model and predicts response values 
for the validation set.
}
