% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runPHATE.R
\name{runPHATE}
\alias{runPHATE}
\title{Run PHATE regression for cross-validation.}
\usage{
runPHATE(train_y, train_x, valid_x, y_factor)
}
\arguments{
\item{train_y}{a vector of numeric values; the response (training set)}

\item{train_x}{a matrix or data frame of numeric values; the data (training set)}

\item{valid_x}{a matrix or data frame of numeric values; the data (validation set)}

\item{y_factor}{a boolean flag; whether the response is categorical or not}
}
\value{
a list containing the training model and the predicted values of the response
using \code{valid_x} and \code{train_x}
}
\description{
Creates a regression model on the embeddings of the data
set found from PHATE and predicts response values for the validation set.
}
