% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runSims.R
\name{runSims}
\alias{runSims}
\title{Run simulations.}
\usage{
runSims(N, P, K, sK, sI, spec, niter, fdr, f_size, ncore, out_path, seed)
}
\arguments{
\item{N}{a single numeric value or a vector; sample size}

\item{P}{a single numeric value or a vector; number of features}

\item{K}{a single numeric value or a vector; number of clusters}

\item{sK}{a single numeric value or a vector; number of significant clusters used to
generate \eqn{Y}}

\item{sI}{a single numeric value or a vector; number of significant interaction terms used
to generate \eqn{Y}}

\item{spec}{a numeric constant; the proportion of iterations that a variable must be selected in to be considered significant}

\item{niter}{an integer; the number of times to run knockoffs}

\item{fdr}{a numeric constant between 0 and 1; the target false discovery rate for knockoffs}

\item{f_size}{an integer; the target size for the subset (number of columns in each subset)}

\item{ncore}{an integer; the number of cores to use for parallel computing}

\item{out_path}{a string path; where to save output}
}
\value{
none
}
\description{
Given different parameter sets, run simulations and save the results for plotting.
Of \code{N}, \code{P}, \code{K}, \code{sK}, and \code{sI}, one should be a vector of values to
loop over and the rest should be single values. This way the varying values of a single parameter
can be plotted on the x-axis against the model MSEs on the y-axis for the simulation line plots.
}
