% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runSLIDE.R
\name{runSLIDE}
\alias{runSLIDE}
\title{running the SLIDE function}
\usage{
runSLIDE(
  y,
  y_path = NULL,
  z_path = NULL,
  z_matrix,
  all_latent_factors,
  lf_path,
  method = 4,
  do_interacts = TRUE,
  fdr = 0.1,
  niter = 500,
  spec = 0.1,
  f_size = NULL
)
}
\arguments{
\item{y}{a matrix of the response vector}

\item{y_path}{a string that points to the response vector}

\item{z_path}{a string that points to the z matrix, user can also input the matrix directly.}

\item{z_matrix}{the z matrix, user can also input a path to the csv file.}

\item{all_latent_factors}{a object contain all latent factors from getLatentFactors.}

\item{lf_path}{a string that points to the all_latent_factors as an RDS object.}

\item{method}{which method of SLIDE to run}

\item{do_interacts}{whether get interaction terms or not}

\item{fdr}{the fdr threshold}

\item{niter}{the number of iterations SLIDE run}

\item{spec}{the threshold to use to choose the latent factors}

\item{f_size}{the number of latent factors to split.}
}
\value{
A list that contains all key information from SLIDE.
}
\description{
Run the SLIDE function and formulate the results and the key parameters into a object.
}
