% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runSLIDEBeta.R
\name{runSLIDEBeta}
\alias{runSLIDEBeta}
\title{Run SLIDE for cross-validation.}
\usage{
runSLIDEBeta(
  train_y,
  valid_y,
  train_z,
  valid_z,
  method,
  spec,
  niter,
  fdr,
  f_size,
  betas,
  top_prop,
  marginals,
  parallel,
  ncore,
  y_factor,
  do_interacts = TRUE
)
}
\arguments{
\item{train_y}{a vector of numeric values; the response (training set)}

\item{valid_y}{a vector of numeric values; the response (validation set)}

\item{train_z}{a matrix or data frame of numeric values; the latent factor data (training set)}

\item{valid_z}{a matrix or data frame of numeric values; the latent factor data (validation set)}

\item{method}{an integer (1, 2, 3, 4); the selection method for determining marginal/interaction significance}

\item{spec}{a numeric constant; the proportion of iterations that a variable must be selected in to be considered significant}

\item{niter}{an integer; the number of times to run knockoffs}

\item{fdr}{a numeric constant between 0 and 1; the target false discovery rate for knockoffs}

\item{f_size}{an integer; the target size for the subset (number of columns in each subset)}

\item{betas}{a vector of numeric values; the coefficients corresponding to the variables in \code{z} (found using Essential Regression)}

\item{top_prop}{a numeric constant between 0 and 1; the proportion of \code{betas} to consider significant}

\item{marginals}{a vector of integers; the indices of the variables to consider significant}

\item{parallel}{a boolean flag; run iterations in parallel or sequentially}

\item{ncore}{an integer; the number of cores to use for parallel computing}

\item{y_factor}{a boolean flag; whether the response is categorical or not}
}
\value{
a list containing the training model and the predicted values of the response
using \code{valid_x}
}
\description{
Creates a regression model from the results of SLIDE and
predicts response values from the validation data.
}
