% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/secondKO.R
\name{secondKO}
\alias{secondKO}
\title{Run aggregated second order knockoffs.}
\usage{
secondKO(z, y, niter = 100, fdr = 0.2, parallel = TRUE)
}
\arguments{
\item{z}{a matrix or data frame of measured data of dimensions \eqn{n} by \eqn{p}}

\item{y}{a matrix or data frame of responses of dimensions \eqn{n} by 1}

\item{niter}{a numeric constant; the number of times to run knockoffs}

\item{fdr}{a numeric constant; the target false discovery rate for knockoffs}

\item{parallel}{a boolean flag; run iterations in parallel or sequentially}
}
\value{
a list including the names of the selected columns of \code{z} that are determined to be significant and a table of their selection frequencies
}
\description{
Iteratively run knockoffs and report variable selection frequencies.
}
