% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectFrequent.R
\name{selectFrequent}
\alias{selectFrequent}
\title{Find marginal variables using Method 2.}
\usage{
selectFrequent(z, y, niter = 1000, fdr = 0.1, parallel = TRUE)
}
\arguments{
\item{z}{a matrix or data frame of measured data of dimensions \eqn{n} by \eqn{p}}

\item{y}{a matrix or data frame of responses of dimensions \eqn{n} by 1}

\item{niter}{a numeric constant; the number of times to run knockoffs}

\item{fdr}{a numeric constant; the target false discovery rate for knockoffs}

\item{parallel}{a boolean flag; run iterations in parallel or sequentially}
}
\value{
a vector of names corresponding to the selected columns of \code{z} that are determined to be significant
}
\description{
Find significant marginal variables:
2 - Significant marginal variable is the single most freqently appearing variable returned from \code{niter} iterations of
    second order knockoffs
}
