% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectLongest.R
\name{selectLongest}
\alias{selectLongest}
\title{Find marginal variables using Method 1.}
\usage{
selectLongest(z, y, niter = 1000, fdr = 0.1)
}
\arguments{
\item{z}{a matrix or data frame of measured data of dimensions \eqn{n} by \eqn{p}}

\item{y}{a matrix or data frame of responses of dimensions \eqn{n} by 1}

\item{niter}{a numeric constant; the number of times to run knockoffs}

\item{fdr}{a numeric constant; the target false discovery rate for knockoffs}
}
\value{
a vector of names corresponding to the selected columns of \code{z} that are determined to be significant
}
\description{
Find significant marginal variables:
1 - Significant marginal variables are those appearing in the longest list returned from \code{niter} iterations of
Gaussian knockoffs.
}
