% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectShortFreq.R
\name{selectShortFreq}
\alias{selectShortFreq}
\title{Find significant variables according to length and frequency.}
\usage{
selectShortFreq(
  z,
  y,
  spec = 0.3,
  fdr = 0.1,
  elbow = FALSE,
  niter = 1000,
  f_size = 100,
  parallel = TRUE
)
}
\arguments{
\item{z}{a matrix or data frame of measured data of dimensions \eqn{n} by \eqn{p}}

\item{y}{a matrix or data frame of responses of dimensions \eqn{n} by 1}

\item{spec}{a numeric constant; the proportion of iterations that a variable must be selected in to be considered significant}

\item{fdr}{a numeric constant; the target false discovery rate for knockoffs}

\item{elbow}{a boolean flag; whether to select significant variables according to the "joint" of the elbow in a histogram or not}

\item{niter}{a numeric constant; the number of times to run knockoffs}

\item{f_size}{a numeric constant; the target size for the subset (number of columns in each subset)}

\item{parallel}{a boolean flag; run iterations in parallel or sequentially}
}
\value{
a vector of names corresponding to the selected columns of \code{z} that are determined to be significant
}
\description{
Significant marginal variables are those appearing in the best list returned from \code{niter} iterations of
second order knockoffs. In this case, "best" refers to the shortest list that has the maximal overlap with the list of
most frequently selected variables. The frequency list is comprised of those variables appearing at least \code{spec}
proportion of the interactions.
}
