% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigBetas.R
\name{sigBetas}
\alias{sigBetas}
\title{Significant \eqn{\beta}s.}
\usage{
sigBetas(betas, cutoff)
}
\arguments{
\item{betas}{a vector of estimates for \eqn{\beta}s}

\item{cutoff}{a numeric constant (proportion) for thresholding (default = 0.10)}
}
\value{
a list of beta indices organized by sign
}
\description{
Find top \eqn{\beta}s according to cutoff percentage.
}
