% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signPerm.R
\name{signPerm}
\alias{signPerm}
\title{Sign Permutation.}
\usage{
signPerm(n)
}
\arguments{
\item{n}{an integer}
}
\value{
matrix of dimensions \eqn{2^n \times n} containing all sign permutations
}
\description{
For a given integer \eqn{n}, generate its sign permutation.
}
