% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singleton.R
\name{singleton}
\alias{singleton}
\title{Check Singletone Node.}
\usage{
singleton(pure_list)
}
\arguments{
\item{pure_list}{estimated indices of pure nodes}
}
\value{
true or false depending upon existence of singleton element
}
\description{
Check if there exists an element of the given list has length equal to 1.
}
