% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SLIDE.R
\name{SLIDE}
\alias{SLIDE}
\title{Run SLIDE.}
\usage{
SLIDE(
  z,
  y,
  method = 4,
  do_interacts = TRUE,
  betas = NULL,
  top_prop = NULL,
  marginals = NULL,
  spec = 0.3,
  fdr = 0.1,
  niter = 1000,
  elbow = FALSE,
  f_size = 100,
  parallel = TRUE,
  ncore = 10
)
}
\arguments{
\item{z}{a matrix or data frame of measured data of dimensions \eqn{n} by \eqn{p}}

\item{y}{a matrix or data frame of responses of dimensions \eqn{n} by 1}

\item{method}{an integer between 1 and 6; the selected method to use}

\item{do_interacts}{a boolean flag; whether to select interaction terms or not}

\item{betas}{a vector of numeric values; the coefficients corresponding to the variables in \code{z} (found using Essential Regression)}

\item{top_prop}{a numeric constant between 0 and 1; the proportion of \code{betas} to consider significant}

\item{marginals}{a vector of integers; the indices of the variables to consider significant}

\item{spec}{a numeric constant; the proportion of iterations that a variable must be selected in to be considered significant}

\item{fdr}{a numeric constant between 0 and 1; the target false discovery rate for knockoffs}

\item{niter}{an integer; the number of times to run knockoffs}

\item{elbow}{a boolean flag; whether to select significant variables according to the "joint" of the elbow in a histogram or not}

\item{f_size}{an integer; the target size for the subset (number of columns in each subset)}

\item{parallel}{a boolean flag; run iterations in parallel or sequentially}

\item{ncore}{an integer; the number of cores to use for parallel computing}
}
\description{
Run SLIDE.
}
