% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slideBCV.R
\name{slide_BCV}
\alias{slide_BCV}
\title{Selects binary structure for SLIDE model using bi-cross-validation}
\usage{
slide_BCV(X, pvec, structure_list, n_fold = 3, p_fold = 3, k_max = 1000,
  eps = 1e-06, center = T)
}
\arguments{
\item{X}{A n x p concatenated data matrix of views X_1,...,X_d.}

\item{pvec}{A vector of values p_1,....,p_d corresponding to the number of measurements within each data view.}

\item{structure_list}{A list of distinct binary structures which forms a candidate set for SLIDE model.}

\item{n_fold}{A number of folds for rows, the default value is 3.}

\item{p_fold}{A number of folds for columns, the default value is 3.}

\item{k_max}{A maximal number of allowed iterations for SLIDE model fitting, the default value is 1000.}

\item{eps}{A convergence tolerance criterion, the default value is 1e-6.}

\item{center}{A logical indicator of whether the data is centered, the default value is TRUE.}
}
\value{
A list with the elements
  \item{structure_list}{A list of candidate structures, same as the input \code{structure_list}.}
  \item{structure_min}{A binary matrix corresponding to the structure with minimal error across the folds.}
  \item{error_mean}{Total error values across the folds for each structure from the list.}
  \item{error}{Matrix of error values for each fold and each structure.}
}
\description{
Selects binary structure for SLIDE model using bi-cross-validation based on the supplied set of candidate structures
}
\examples{
n = 100
p1 = 25
p2 = 25
data = generateModel1(n = n, pvec = c(p1, p2))

# Standardize the data
out = standardizeX(data$X, pvec = c(p1,p2))

# Create a list of candidate structures
out_struct <- create_structure_list(out$X, data$pvec, lambda_max = max(out$svec),
 standardized = TRUE)

# Select a structure from the list using bcv procedure
out_bcv <- slide_BCV(out$X, data$pvec, structure_list = out_struct$Slist)
out_bcv$structure_min
}
