% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SLIDEcv.R
\name{SLIDEcv}
\alias{SLIDEcv}
\title{Cross validation for SLIDE latent factors}
\usage{
SLIDEcv(yaml_path = NULL, nrep = 20, k = 10)
}
\arguments{
\item{yaml_path}{the string to a yaml path where optimizeSLIDE has been run; can either be the original yaml file (to run CV on all runs),
or the yaml for a specific run}

\item{nrep}{number of replicates for cross validation (default 20)}

\item{k}{number of folds for cross validation (default 10)}
}
\description{
Cross validation for SLIDE latent factors
}
