% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Algorithm1Cwrapper.R
\name{solve_optim}
\alias{solve_optim}
\title{Solve optimization algorithm for one value of lambda}
\usage{
solve_optim(X, lambda, pvec, k_max = 1000, eps = 1e-06,
  Ustart = NULL)
}
\arguments{
\item{X}{}

\item{lambda}{}

\item{pvec}{}

\item{k_max}{}

\item{eps}{}

\item{Ustart}{}
}
\description{
Solve optimization algorithm for one value of lambda
}
