% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solveRow.R
\name{solveRow}
\alias{solveRow}
\title{???}
\usage{
solveRow(col_ind, C, lambda)
}
\arguments{
\item{col_ind}{indices of columns}

\item{C}{matrix?}

\item{lambda}{\eqn{\labmda}}
}
\value{
value
}
\description{
???
}
