% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitSecond.R
\name{splitSecond}
\alias{splitSecond}
\title{Run aggregated second order knockoffs with splitting.}
\usage{
splitSecond(
  z,
  y,
  spec = 0.1,
  fdr = 0.05,
  f_size = 100,
  niter = 20,
  parallel = T
)
}
\arguments{
\item{z}{a matrix or data frame of measured data of dimensions \eqn{n} by \eqn{p}}

\item{y}{a matrix or data frame of responses of dimensions \eqn{n} by 1}

\item{spec}{a numeric constant; the maximal difference in frequency proportions}

\item{fdr}{a numeric constant; the target false discovery rate for knockoffs}

\item{f_size}{the subset size (number of columns); must be less than or equal to \eqn{p}}

\item{niter}{a numeric constant; the number of times to run knockoffs}

\item{parallel}{a boolean flag; run iterations in parallel or sequentially}
}
\value{
a vector of names corresponding to the selected columns of \code{z} that are determined to be significant
}
\description{
Split data into subsets and run knockoffs on each subset separately. Aggregate results and do one final run of knockoffs to get final selected variables.
}
