% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Standardization.R
\name{standardizeX}
\alias{standardizeX}
\title{Standardization of multi-view data.}
\usage{
standardizeX(X, pvec, center = T)
}
\arguments{
\item{X}{A n x p concatenated data matrix of views X_1,...,X_d.}

\item{pvec}{A vector of values p_1,....,p_d corresponding to the number of measurements within each  view.}

\item{center}{A logical indicator of whether the columns of X should be centered. The default value is TRUE.}
}
\value{
A list with the elements
  \item{X}{A n x p concatenated data matrix that has been column-centered and standardized so that Frobenius norm within each view is equal to one.}
  \item{svec}{A vector of largest singular values for each view after centering and standardization.}
  \item{norms}{A vector of Frobenius norms for each view before scaling.}
  \item{Xmean}{A vector of column means of X before centering. A zero vector is returned if \code{center = F}.}
}
\description{
Performs column-centering and standardization so that the Frobenius norm within each view is equal to one.
}
\examples{
n = 100
p1 = 40
p2 = 60
X1 = matrix(rnorm(n*p1), n, p1)
X2 = matrix(rnorm(n*p2), n, p2)
X = cbind(X1, X2)
out = standardizeX(X, pvec = c(p1,p2))
}
