% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testUpsilon.R
\name{testUpsilon}
\alias{testUpsilon}
\title{Find significant upsilons.}
\usage{
testUpsilon(
  upsilon,
  y,
  spec = 0.1,
  fdr = 0.1,
  niter = 1000,
  f_size = 100,
  parallel = TRUE,
  elbow = FALSE
)
}
\arguments{
\item{upsilon}{a data frame; the values of upsilon}

\item{y}{a vector; response values}

\item{spec}{a numeric constant; the proportion of iterations that a variable must be selected in to be considered significant}

\item{fdr}{a numeric constant; the target false discovery rate for knockoffs}

\item{niter}{a numeric constant; the number of times to run knockoffs}

\item{f_size}{a numeric constant; the target size for the subset (number of columns in each subset)}

\item{parallel}{a boolean flag; run iterations in parallel or sequentially}
}
\value{
a data frame
}
\description{
Use second order knockoffs to find the significant upsilons according to length and frequency.
}
