% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threshA.R
\name{threshA}
\alias{threshA}
\title{Threshold \eqn{A}.}
\usage{
threshA(A, mu, scale = FALSE)
}
\arguments{
\item{A}{an estimated matrix}

\item{mu}{threshold, a positive value}

\item{scale}{boolean for whether to normalize}
}
\value{
thresholded version of matrix \code{A}
}
\description{
Threshold the given matrix \code{A} based on the given \eqn{\mu}. If \code{scale} is true,
then normalize each row of \code{A} such that the \eqn{L_1} norm of each row is not larger than 1.
}
