% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toCont.R
\name{toCont}
\alias{toCont}
\title{Categorical To Continuous Data.}
\usage{
toCont(y, order = NULL)
}
\arguments{
\item{y}{a response vector of dimension \eqn{n} of categorical values}

\item{order}{a vector indicating the level ordering of \eqn{y}}
}
\value{
a list containing the old (categorical) \eqn{y} and the new (continuous) \eqn{y}
}
\description{
Transform categorical response vector into continuous values.
}
