% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slide_data.R
\docType{data}
\name{UN_sig}
\alias{UN_sig}
\title{Protein Expression Levels in an Uninfected Cell Population}
\format{An object of class \code{data.frame} with 781 rows and 7 columns.}
\usage{
UN_sig
}
\description{
Simulated data from an experiment measuring viral reprogramming of lymphocyte surface proteins. 
These cells represent the uninfected cell population to which the infected cells will be compared to.
}
\keyword{datasets}
