% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VarExplained.R
\name{var_explained}
\alias{var_explained}
\title{Percent of variance explained by each component of fitted SLIDE model}
\usage{
var_explained(slide, pvec)
}
\arguments{
\item{slide}{Output from \code{slide} function}

\item{pvec}{A vector of values p_1,....,p_d corresponding to the number of measurements within each  view.}
}
\value{
A list with the elements
  \item{var_total}{A vector of length r (number of components in fitted SLIDE model), each element is a percentage variance explained in the total concatenated dataset by that component}
  \item{var_vew}{A d x r matrix, each element is a percentage variance explained in the respective view by that component.}
}
\description{
Percent of variance explained by each component of fitted SLIDE model
}
\examples{
n = 25
# Two matched datasets
p1 = p2 = 10
data = generateModel1(n = n, pvec = c(p1, p2))
out_slide = slide(X = data$X, pvec = c(p1, p2))
var_explained(out_slide, pvec = c(p1, p2))

}
