% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcControlPerformance.R
\name{calcControlPerformance}
\alias{calcControlPerformance}
\title{calcControlPerformance}
\usage{
calcControlPerformance(z_matrix, y, SLIDE_res, niter, condition, out_path)
}
\arguments{
\item{z_matrix}{the z_matrix, the output of CalcZMatrix function.}

\item{y}{the response matrix.}

\item{SLIDE_res}{the output of GetTopFeatures function.}

\item{niter}{a integer for number of iterations.}

\item{condition}{a string on whether to run auc or corr.}

\item{out_path}{the output path.}
}
\value{
The z matrix calculated by the CalcZMatrix function
}
\description{
Calculate control model performances including the real model, the model with random interaction terms and full random model.
}
