% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkDataParams.R
\name{checkDataParams}
\alias{checkDataParams}
\title{Check if the input and the yaml files are set correctly.}
\usage{
checkDataParams(input_params, pipeline = NULL)
}
\arguments{
\item{input_params}{a list of the input parameters from the YAML file.}

\item{pipeline}{a integer indicates which step of the pipeline is the YAML file for.}
}
\description{
Check if the input and the yaml files are set correctly.
}
