% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CheckDataParams.R
\name{CheckDataParams}
\alias{CheckDataParams}
\title{Check if the input and the yaml files are set correctly.}
\usage{
CheckDataParams(yaml_path, x_path, y_path, pipeline)
}
\arguments{
\item{yaml_path}{a string that points to the yaml file.}

\item{x_path}{a string that points to the x matrix.}

\item{y_path}{a string that points to the y vector.}

\item{pipeline}{a integer indicates which step of the pipeline is the yaml file for.}
}
\description{
Check if the input and the yaml files are set correctly.
}
