% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetTopFeatures.R
\name{GetTopFeatures}
\alias{GetTopFeatures}
\title{Output the top features in each latent factors}
\usage{
GetTopFeatures(
  x_path,
  y_path,
  er_path,
  out_path,
  SLIDE_res,
  num_top_feats = 10,
  condition
)
}
\arguments{
\item{x_path}{a string that points to the x matrix.}

\item{y_path}{a string that points to the y vector.}

\item{er_path}{a string that points to the final er result as an RDS object.}

\item{out_path}{a string that points to a folder where txt files of each latent factors will be outputted.}

\item{SLIDE_res}{the object outputted by the runSLIDE function.}

\item{num_top_feats}{the number of top features to get from each criteria such as A loadings or AUCs.}

\item{condition}{either "corr" or "auc", indicates which one to use to get top features.}
}
\value{
The SLIDE_res object with feature results added to one of the slots.
}
\description{
Output the top features in each latent factors by taking a union of features with highest A loadings and spearman correlations or univariate AUCs
}
