% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runSLIDE.R
\name{runSLIDE}
\alias{runSLIDE}
\title{running the SLIDE function}
\usage{
runSLIDE(
  y_path,
  z_path = NULL,
  z_matrix,
  er_path,
  method = 4,
  do_interacts = TRUE,
  fdr = 0.1,
  niter = 500,
  spec = 0.1
)
}
\arguments{
\item{y_path}{a string that points to the y vector}

\item{z_path}{a string that points to the z matrix, user can also input the matrix directly.}

\item{z_matrix}{the z matrix, user can also input a path to the csv file.}

\item{er_path}{a string that points to the final er result as an RDS object.}

\item{method}{which method of SLIDE to run}

\item{do_interacts}{whether get interaction terms or not}

\item{fdr}{the fdr threshold}

\item{niter}{the number of iterations SLIDE run}

\item{spec}{the threshold to use to choose the latent factors}
}
\value{
A list that contains all key information from SLIDE.
}
\description{
Run the SLIDE function and formulate the results and the key parameters into a object.
}
