% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampleCV.R
\name{sampleCV}
\alias{sampleCV}
\title{Sample Cross Validation.}
\usage{
sampleCV(
  y,
  z_matrix,
  SLIDE_res,
  fraction = 2/3,
  condition,
  sampleCV_iter = 100,
  logistic = FALSE,
  out_path
)
}
\arguments{
\item{y}{the response matrix.}

\item{z_matrix}{the z_matrix, the output of CalcZMatrix function}

\item{SLIDE_res}{the output of GetTopFeatures function.}

\item{fraction}{the fraction of samples to choose for sampling.}

\item{condition}{use auc or corr(spearman) to evaluate performance.}

\item{sampleCV_iter}{a integer for number of iterations.}

\item{logistic}{a boolean flag on wether running logistic or lm.}

\item{out_path}{the output path.}
}
\value{
the mean performance.
}
\description{
Calculate SLIDE performance using the sample CV framework for a quick evaluation. This is meant to help user to choose which model to use for the more rigorous cross validation.
}
