% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zeroFiltering.R
\name{zeroFiltering}
\alias{zeroFiltering}
\title{Reduce sparsity of the data matrix by deleting samples and features that are expressed in few features and samples.}
\usage{
zeroFiltering(x, y, col_thresh, row_thresh)
}
\arguments{
\item{col_thresh}{filtered out columns with number of zeros greater than col_thresh, if entered as nrow of x, no columns will be deleted.}

\item{row_thresh}{filtered out rows with number of zeros greater than row_thresh, if entered as ncol of x, no rows will be deleted.}

\item{unfiltered_x}{the unfiltered x matrix}

\item{unfiltered_y}{the unfiltered y vector}
}
\value{
a list including the filtered x matrix and y vector
}
\description{
Reduce sparsity of the data matrix by deleting samples and features that are expressed in few features and samples.
}
