% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ZeroFiltering.R
\name{ZeroFiltering}
\alias{ZeroFiltering}
\title{Reduce sparsity of the data matrix by deleting samples and features that are expressed in few features and samples.}
\usage{
ZeroFiltering(data, g_thresh, c_thresh)
}
\arguments{
\item{data}{the data matrix}

\item{g_thresh}{samples with number of zeros greater then g_thresh will be filtered out}

\item{c_thresh}{features with number of zeros greater then g_thresh will be filtered out}
}
\description{
Reduce sparsity of the data matrix by deleting samples and features that are expressed in few features and samples.
}
